+++ lib/hyperglot/orthography.py
@@ -1,7 +1,10 @@
-from typing import List, Set
-import unicodedata2
+import os
+import yaml
 import logging
-
+import unicodedata2
+from typing import List, Set
+from functools import lru_cache
+from hyperglot import DB_EXTRA
 from hyperglot.shaper import Shaper
 from hyperglot.parse import (
     parse_chars,
@@ -14,6 +17,19 @@ from hyperglot.parse import (
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.WARNING)
+
+
+@lru_cache
+def get_scripts():
+    with open(os.path.join(DB_EXTRA, "script-names.yaml"), "rb") as f:
+        return yaml.load(f, Loader=yaml.Loader)
+
+
+def get_script_iso(name: str) -> str:
+    scripts = get_scripts()
+    if name not in scripts:
+        raise NotImplementedError(f"Missing script name to ISO mapping for {name}")
+    return scripts[name]["iso_15924"]
 
 
 def is_mark(c):
@@ -41,9 +57,12 @@ class Orthography(dict):
 
     defaults = {"preferred_as_group": False}
 
-    def __init__(self, data: dict):
+    def __init__(self, data: dict, include_script_iso: bool = False):
         self.update(self.defaults)
         self.update(data)
+
+        if include_script_iso:
+            self["script_iso"] = get_script_iso(self["script"])
 
     @property
     def presentation(self):