+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - DATA: Use "Coptic" instead of "Coptic/Numbian" script name
 - DATA: Use "Burmese" script for language "Mon"
 - DATA: Use "Baybayin" script name instead of "Tagalog (Baybayin, Alibata)"
+- TWEAK: Added `include_script_iso` flag to `hyperglot.orthography.Orthography` constructor to return the ISO 15924 script code
+- DATA: Added `lib/extra_data/script-names.yaml` with a list of all current Hyperglot scripts and a mapping to their ISO 15924 code equivalent
 
 ## 0.6.3 (08.04.2024)
 - FIX: Set correct default values for `Language.status` and `Orthography.preferred_as_group` and provide validation and tests for these.